Rename person details package and classes to person account
